+++ src/Game/Game/Segment_29f3.cs
@@ -143,7 +143,7 @@ namespace OpenCiv1
 			this.oCPU.AX.Word = (ushort)this.oGameData.HumanPlayerID;
 			this.oCPU.CMP_UInt16((ushort)playerID2, this.oCPU.AX.Word);
 			if (this.oCPU.Flags.NE) goto L0130;
-			this.oCPU.CX.Word = (ushort)this.oGameData.DifficultyLevel;
+			this.oCPU.CX.Word = (ushort)((short)this.oGameData.DifficultyLevel);
 			this.oCPU.CX.Word = this.oCPU.INC_UInt16(this.oCPU.CX.Word);
 			this.oCPU.AX.Word = this.oCPU.ReadUInt16(this.oCPU.SS.Word, (ushort)(this.oCPU.BP.Word - 0x10));
 			this.oCPU.IMUL_UInt16(this.oCPU.AX, this.oCPU.DX, this.oCPU.CX.Word);
